# registry/api-clients/api;/api-clients/api/api/replacer.bat/model/software.ts
/**
 * Registry-Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Software { 
    id?: number;
    license?: string | null;
    name?: string | null;
    description?: string | null;
    price?: number;
    productLink?: string | null;
    userId?: string | null;
    isActive?: boolean;
}