# registry/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Registry-Backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Softwares/Single/{id}": {
      "get": {
        "tags": [
          "Softwares"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Software"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Software"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Software"
                }
              }
            }
          }
        }
      }
    },
    "/api/Softwares/Delete/{id}": {
      "delete": {
        "tags": [
          "Softwares"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Softwares/UpdateSoftware": {
      "patch": {
        "tags": [
          "Softwares"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Software"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Software"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Software"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Softwares/AddSoftware": {
      "post": {
        "tags": [
          "Softwares"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Software"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Software"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Software"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Software": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "license": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "productLink": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}